feat: NormalSettingItem 컴포넌트 분리
